Change path to be derived, fix smoke test
